[ENGOPS-585] cda-core compile fails without ehache
This module has explicit dependencies on ehcache, but is expecting it to get pulled in transitively. This does not happen in the new nightly builds due to a fresh ivy cache and the more explicit pentaho artifact versioning that happens there.

Also removed attempt at module wide exclude of ehache.